供应商工程商

+++ src/main/java/com/zhuhuibao/mybatis/memberReg/mapper/MemberRegMapper.java
@@ -29,6 +29,6 @@ public interface MemberRegMapper {
 	//更新找回密码邮箱验证是否通过
 	int updateMemberValidatePass(Member member);
 	
-	Member findMemberInfoById(Integer id);
+	Member findMemberInfoById(Long id);
 	
 }
